mejora codigo
