+++ src/main/java/co/edu/unicauca/parcial/services/services/asignatura/IAsignturaService.java
@@ -1,4 +1,4 @@
-package co.edu.unicauca.parcial.services.services.AsignaturaService;
+package co.edu.unicauca.parcial.services.services.asignatura;
 
 import co.edu.unicauca.parcial.services.DTO.AsignaturaDTO;
