+++ src/main/java/co/edu/unicauca/parcial/services/services/curso/CursoServiceImpl.java
@@ -1,4 +1,4 @@
-package co.edu.unicauca.parcial.services.services.CursoService;
+package co.edu.unicauca.parcial.services.services.curso;
 
 import java.util.List;
 
@@ -26,17 +26,14 @@ public class CursoServiceImpl implements ICursoService {
     @Override
     @Transactional(readOnly = true)
     public List<CursoDTO> getAllCursos() {
-        System.out.println("Invocando a listar cursos");
         Iterable<Curso> cursos = cursoRepository.findAll();
-        List<CursoDTO> cursosDTO = modelMapper.map(cursos, new TypeToken<List<CursoDTO>>() {
+        return modelMapper.map(cursos, new TypeToken<List<CursoDTO>>() {
         }.getType());
-        return cursosDTO;
     }
 
     @Override
     @Transactional()
     public CursoDTO createCurso(CursoDTO curso) {
-        System.out.println("Invocando al metodo crear curso");
         Curso objCurso = modelMapper.map(curso, Curso.class);
         CursoDTO cursoDTO = null;
         Curso cursoAlmacenado = cursoRepository.save(objCurso);
@@ -49,16 +46,13 @@ public class CursoServiceImpl implements ICursoService {
     @Override
     @Transactional(readOnly = true)
     public CursoDTO getCursoById(Integer id) {
-        System.out.println("Invocando al metodo buscar curso por id: " + id);
         Curso curso = cursoRepository.findById(id).orElse(null);
-        CursoDTO cursoDTO = modelMapper.map(curso, CursoDTO.class);
-        return cursoDTO;
+        return modelMapper.map(curso, CursoDTO.class);
     }
 
     @Override
     @Transactional(readOnly = false)
     public CursoDTO updateCurso(Integer id, CursoDTO curso) {
-        System.out.println("Invocando al metodo actualizar curso");
         Curso objCurso = cursoRepository.findById(id).orElse(null);
         CursoDTO cursoDTO = null;
         if (objCurso != null) {
@@ -75,7 +69,6 @@ public class CursoServiceImpl implements ICursoService {
     @Override
     @Transactional(readOnly = false)
     public boolean deleteCurso(Integer id) {
-        System.out.println("Invocando al metodo eliminar curso por id: " + id);
         boolean result = false;
         Curso curso = cursoRepository.findById(id).orElse(null);
         if (curso != null) {
